# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Assemble Yourself",
        "description": "Приложение для личностного роста",
        "contact": {
            "email": "v.a.kuznetsov-83@yandex.ru"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/courses": {
            "get": {
                "tags": [
                    "courses"
                ],
                "summary": "Получение курсов",
                "description": "Получает все курсы содержащиеся в базе",
                "operationId": "getCourses",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CourseModel"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "courses"
                ],
                "summary": "Создание курса",
                "description": "Создает курс с заданными параметрами",
                "operationId": "createCourse",
                "requestBody": {
                    "$ref": "#/components/requestBodies/CourseArray"
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CourseModel"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Переданы не все обязательные поля",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Поле id обязательно для заполнения"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Курс может составлять только авторизованный пользователь"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/courses/{id}": {
            "get": {
                "tags": [
                    "courses"
                ],
                "summary": "Получение курса",
                "description": "Получает курсы по ID курса",
                "operationId": "getCourseById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID по которому выберается курс",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CourseModel"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Такого курса не существует",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Course not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "courses"
                ],
                "summary": "Изменение курса",
                "description": "Изменяет курс с заданными параметрами",
                "operationId": "deleteCourse",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID по которому удаляется курс",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Курс может удалить только авторизованный пользователь"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Курса не существует",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Такого курса не существует"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Пользователь не автор курса",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Курс может удалить только его создатель"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "courses"
                ],
                "summary": "Изменение курса",
                "description": "Изменяет курс с заданными параметрами",
                "operationId": "updateCourse",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID по которому выберается курс",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/CourseArray"
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CourseModel"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Переданы не все обязательные поля",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Поле id обязательно для заполнения"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Курс может менять только авторизованный пользователь"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Курса не существует",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Такого курса не существует"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Пользователь не автор курса",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Описание курса может менять только автор курса"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CourseModel": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "course_name": {
                        "description": "Название курса",
                        "type": "string",
                        "readOnly": "true"
                    },
                    "description": {
                        "description": "Описание курса",
                        "type": "string",
                        "readOnly": "true"
                    },
                    "tag": {
                        "type": "string",
                        "readOnly": "true"
                    },
                    "cover_url": {
                        "description": "URL адрес обложки",
                        "type": "string"
                    },
                    "author": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "start_date": {
                        "description": "Дата начала курса",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "end_date": {
                        "description": "Дата окончания курса",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "course_program": {
                        "description": "Содержание программы курса",
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "created_at": {
                        "description": "Initial creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "deleted_at": {
                        "description": "Soft delete timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Courses"
                }
            },
...
        },
        "requestBodies": {
            "CourseArray": {
                "description": "Перечень свойств курса",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "course_name": {
                                    "description": "Название курса",
                                    "type": "string"
                                },
                                "description": {
                                    "description": "Описание курса",
                                    "type": "string"
                                },
                                "tag": {
                                    "type": "string"
                                },
                                "cover_url": {
                                    "description": "URL адрес обложки",
                                    "type": "string"
                                },
                                "start_date": {
                                    "description": "Дата начала курса",
                                    "type": "string",
                                    "format": "date-time"
                                },
                                "end_at": {
                                    "description": "Дата окончания курса",
                                    "type": "string",
                                    "format": "date-time"
                                },
                                "course_program": {
                                    "description": "Содержание программы курса",
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "scheme": "bearer",
                "description": "Enter token in format (Bearer {Token})",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}